pyproject.toml
[project]
name = "my_project"
version = "0.1.0"
description = "A simple Python project"

requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.2",
    "ngrok>=1.5.1",
    "pipecat-ai-flows>=0.0.21",
    "pipecat-ai[cartesia,deepgram,local-smart-turn-v3,openai,openrouter,runner,silero,webrtc,websocket]>=0.0.85",
    "pipecatcloud>=0.2.4",
    "pyngrok>=7.4.0",
    "requests>=2.0.0",
    "twilio>=9.8.3",
]
